more changes for 64 bits.  the moral of the story:  only use long if you mean it.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@3845 23e1bc29-d954-0410-9766-47de34158ba0

+++ Solvers/component/SolverPartialPlanWriter.cc
@@ -471,7 +471,7 @@ namespace EUROPA {
 	numTokens = numVariables = numConstraints = 0;
 
 	char stepstr[NBBY * sizeof(nstep) * 28/93 + 4];
-	sprintf(stepstr, "%d", nstep);
+	sprintf(stepstr, "%d", (int) nstep);
     
 	std::string stepnum(STEP + stepstr);
